requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
backcall=0.1.0=py36_0
blas=1.0=mkl
bleach=3.0.2=py36_0
ca-certificates=2018.12.5=0
certifi=2018.11.29=py36_0
cffi=1.11.5=py36h74b6da3_1
cloudpickle=0.6.1=py36_0
colorama=0.4.1=py36_0
cycler=0.10.0=py36h009560c_0
cytoolz=0.9.0.1=py36hfa6e2cd_1
dask-core=1.0.0=py36_0
decorator=4.3.0=py36_0
entrypoints=0.2.3=py36_2
freetype=2.9.1=ha9979f8_1
hdf5=1.8.20=hac2f561_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
imageio=2.4.1=py36_0
intel-openmp=2019.1=144
ipykernel=5.1.0=py36h39e3cac_0
ipython=7.2.0=py36h39e3cac_0
ipython_genutils=0.2.0=py36h3c5d0ee_0
ipywidgets=7.4.2=py36_0
jedi=0.13.2=py36_0
jinja2=2.10=py36_0
jpeg=9b=hb83a4c4_2
jsonschema=2.6.0=py36h7636477_0
jupyter=1.0.0=py36_7
jupyter_client=5.2.4=py36_0
jupyter_console=6.0.0=py36_0
jupyter_contrib_core=0.3.3=py_2
jupyter_contrib_nbextensions=0.5.0=py36_1000
jupyter_core=4.4.0=py36_0
jupyter_highlight_selected_word=0.2.0=py36_1000
jupyter_latex_envs=1.4.4=py36_1000
jupyter_nbextensions_configurator=0.4.0=py36_1000
kiwisolver=1.0.1=py36h6538335_0
libiconv=1.15=hfa6e2cd_1003
libopencv=3.4.2=h20b85fd_0
libpng=1.6.35=h2a8f88b_0
libsodium=1.0.16=h9d3ae62_0
libtiff=4.0.9=h36446d0_2
libxml2=2.9.8=h9ce36c8_1005
libxslt=1.1.32=heafd4d3_1002
lxml=4.2.5=py36heafd4d3_1000
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.1.0=py36he774522_0
matplotlib=3.0.2=py36hc8f65d3_0
mistune=0.8.4=py36he774522_0
mkl=2019.1=144
mkl_fft=1.0.6=py36h6288b17_0
mkl_random=1.0.2=py36h343c172_0
msys2-conda-epoch=20160418=1
nbconvert=5.3.1=py36_0
nbformat=4.4.0=py36h3a5bc1b_0
networkx=2.2=py36_1
ninja=1.8.2=py36he980bc4_1
notebook=5.7.4=py36_0
numpy=1.15.4=py36h19fb1c0_0
numpy-base=1.15.4=py36hc3f5095_0
olefile=0.46=py36_0
opencv=3.4.2=py36h40b0b35_0
openssl=1.1.1=he774522_0
pandas=0.23.4=py36h830ac7b_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py36_1
parso=0.3.1=py36_0
pickleshare=0.7.5=py36_0
pillow=5.3.0=py36hdc69c19_0
pip=18.1=py36_0
prometheus_client=0.5.0=py36_0
prompt_toolkit=2.0.7=py36_0
py-opencv=3.4.2=py36hc319ecb_0
pycparser=2.19=py36_0
pygments=2.3.1=py36_0
pyparsing=2.3.0=py36_0
pyqt=5.9.2=py36h6538335_2
python=3.6.7=h9f7ef89_2
python-dateutil=2.7.5=py36_0
pytorch=1.0.0=py3.6_cuda90_cudnn7_1
pytz=2018.7=py36_0
pywavelets=1.0.1=py36h8c2d366_0
pywinpty=0.5.5=py36_1000
pyyaml=3.13=py36hfa6e2cd_1001
pyzmq=17.1.2=py36hfa6e2cd_0
qt=5.9.7=vc14h73c81de_0
qtconsole=4.4.3=py36_0
scikit-image=0.14.1=py36ha925a31_0
scikit-learn=0.20.1=py36hb854c30_0
scipy=1.1.0=py36h29ff71c_2
send2trash=1.5.0=py36_0
setuptools=40.6.3=py36_0
sip=4.19.8=py36h6538335_0
six=1.12.0=py36_0
sqlite=3.26.0=he774522_0
terminado=0.8.1=py36_1
testpath=0.4.2=py36_0
tk=8.6.8=hfa6e2cd_0
toolz=0.9.0=py36_0
torchvision=0.2.1=py_2
tornado=5.1.1=py36hfa6e2cd_0
tqdm=4.28.1=py36h28b3542_0
traitlets=4.3.2=py36h096827d_0
vc=14.1=h21ff451_3
vs2015_runtime=15.5.2=3
wcwidth=0.1.7=py36h3d5aa90_0
webencodings=0.5.1=py36_1
wheel=0.32.3=py36_0
widgetsnbextension=3.4.2=py36_0
wincertstore=0.2=py36h7fe50ca_0
winpty=0.4.3=4
yaml=0.1.7=hfa6e2cd_1001
zeromq=4.2.5=he025d50_1
zlib=1.2.11=h62dcd97_3
